requirements.txt
streamlit==1.32.0
python-dotenv==1.1.0
neo4j==5.28.1
openai==1.78.1
tiktoken==0.9.0
langchain==0.3.25
langchain-openai==0.3.18
langchain-community==0.3.24 